+++ src/components/SelectPeriod/PeriodPickerModal.module.css
@@ -1,0 +1,107 @@
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 375px;
+  height: 812px;
+  background: rgba(69, 69, 69, 0.6);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.wrapper {
+  display: flex;
+  width: 288px;
+  height: 290px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 0;
+  border-radius: 12px;
+  background: #f1f1f1;
+}
+
+.contentWrapper {
+  padding: 12px 16px;
+  margin-bottom: 20px;
+}
+.dateButtonWrapper {
+  display: flex;
+  margin-bottom: 20px;
+}
+.dateButtonWrapper .activeButton {
+  /*버튼 배경 css*/
+  display: flex;
+  padding: 6px 32px;
+  justify-content: center;
+  align-items: center;
+  background: #00befc;
+  /*버튼 폰트 css*/
+  color: #fafafa;
+  font-family: Pretendard;
+  font-size: 16px;
+  font-weight: 600;
+}
+.dateButtonWrapper .inactiveButton {
+  /*버튼 배경 css*/
+  display: flex;
+  padding: 6px 32px;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid #00befc;
+  background: #f8f8f8;
+  /*버튼 폰트 css*/
+  color: #00befc;
+  font-family: Pretendard;
+  font-size: 16px;
+  font-weight: 600;
+}
+
+.checkButtonWrapper {
+  display: flex;
+  height: 43px;
+  padding: 10px 0px;
+  border-radius: 0px 0px 12px 12px;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+  align-self: stretch;
+  background: #fafafa;
+}
+.checkButtonWrapper .cancelButton {
+  /*버튼 배경 css*/
+  display: flex;
+  width: 124px;
+  height: 39px;
+  border: 1px solid #ababab;
+  border-radius: 10px;
+  justify-content: center;
+  align-items: center;
+  flex-shrink: 0;
+  background-color: #ffffff;
+  /*버튼 폰트 css*/
+  color: #ababab;
+  font-family: Pretendard;
+  font-size: 16px;
+  font-weight: 700;
+}
+.checkButtonWrapper .confirmButton {
+  /*버튼 배경 css*/
+  display: flex;
+  width: 124px;
+  height: 39px;
+  border-radius: 10px;
+  justify-content: center;
+  align-items: center;
+  flex-shrink: 0;
+  background-color: #00befc;
+  /*버튼 폰트 css*/
+  color: #fff;
+  font-family: Pretendard;
+  font-size: 16px;
+  font-weight: 700;
+}
+/*Picker 실선 제거*/
+.customPicker > div:last-child > div {
+  display: none;
+}
